activated and updated tests for the source readers
